nextjs-app/services/avatar-service/requirements.txt
fastapi
uvicorn[standard]
numpy
scipy
Pillow
pygltflib
trimesh
mediapipe
